committing to C low level langauge

+++ 0x00-hello_world/6-size.c
@@ -2,7 +2,7 @@
 int main(void)
 {
 printf("Size of Char: %2d byte(s) \n", sizeof(char));
-printf("Size of Int: %2d byte(si) \n", sizeof(short int));
+printf("Size of Int: %2d byte(si) \n", sizeof(long int));
 printf("Size of Long Int: %2d byte(s) \n", sizeof(long int));
 printf("Size of long long int: %2d byte(s) \n", sizeof(double));
 printf("Size of Float: %2d byte(s) \n", sizeof(float));
